ai-engine/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "defi-ai-engine"
version = "0.1.0"
description = "AI engine for Nexus Yield optimization"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.10.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "scikit-learn>=1.5.0",
    "scipy>=1.14.0",
    "psycopg2-binary>=2.9.0",
    "redis>=5.2.0",
    "bullmq>=1.0.0",
    "aiohttp>=3.9.0",
    "websockets>=13.0",
    "pyyaml>=6.0.0",
    "asyncpg>=0.29.0",
]

[project.optional-dependencies]
ml = [
    "torch>=2.2.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "mypy>=1.13.0",
]

[tool.setuptools.packages.find]
where = ["src"]
